refactor: stream lib no longer compiled in backend lib, allows optimization
