+++ backend/BUILD.bazel
@@ -13,7 +13,13 @@ rust_library(
     crate_name = "backend_lib",
     srcs = glob(
         ["src/**/*.rs"],
-        exclude = ["src/main.*.rs"]
+        exclude = [
+            "src/bin/",
+            "src/main.*.rs",
+            # Exclude following code that contains egui and eframe.
+            # See ticket #30 for more information.
+            "src/camera/stream/**/*.rs",
+        ]
     ),
     crate_features = ["fake"],
     rustc_flags = select({
@@ -27,7 +33,6 @@ rust_library(
         "@crates//:clap",
         "@crates//:crossbeam",
         "@crates//:rand",
-        "@crates//:eframe",
         "@crates//:serde",
         "@crates//:serde_json",
         "@crates//:tokio",
@@ -35,6 +40,29 @@ rust_library(
         "@crates//:uuid",
     ],
 )
+
+# Define a secondary library for streaming, which uses eframe and egui.
+# We do this because optimized release builds will use the `abort` panic
+# strategy to decrease binary size, but eframe and egui require the `unwind`
+# panic strategy. We would like to exclude this library from optimization
+# since it would throw an error and it's not needed in the final binary anyways.
+# See ticket #30 for more information.
+rust_library(
+    name = "stream_lib",
+    crate_name = "stream_lib",
+    srcs = glob(["src/camera/stream/**/*.rs"]),
+    deps = [
+        ":backend_lib",
+        "@crates//:aravis",
+        "@crates//:crossbeam",
+        "@crates//:eframe",
+        "@crates//:egui",
+    ],
+)
+
+# -----------------------------------------------------------------------------
+# MAIN BINARY
+# -----------------------------------------------------------------------------
 
 # Note, the main reason for splitting into dev and prod
 # binaries is to handle differences in how the applications serve
@@ -46,11 +74,7 @@ rust_library(
 # actually starts up the CV pipelines per camera. The `analyze-throw`
 # route will also call different functions based on this configuration.
 
-# -----------------------------------------------------------------------------
-# MAIN BINARY
-# -----------------------------------------------------------------------------
-
-# Dev mode for the backend doesn't have to be optimized. In this mode,
+# Dev modes for the backend doesn't have to be optimized. In this mode,
 # the backend can serve the API via command line and it runs alongside
 # the Next.js server. Since the Next.js requests come from a different
 # origin, CORS is needed. It will not have any embedded assets.
@@ -79,10 +103,11 @@ rust_binary(
     ],
 )
 
-# Prod mode for the backend should be optimized. In this mode, the 
+# Prod modes for the backend should be optimized. In this mode, the 
 # backend will serve the API but instead of serving the Next.js server,
 # it will embed the frontend's static exports using rust-embed and 
 # serve using axum_embed.
+
 rust_binary(
     name = "prod-fake",
     srcs = ["src/main.prod.rs"],
@@ -192,6 +217,7 @@ rust_binary(
     srcs = ["src/bin/live_stream.rs"],
     deps = [
         ":backend_lib",
+        ":stream_lib",
         "@crates//:clap",
         "@crates//:crossbeam",
         "@crates//:ctrlc",
