+++ backend/src/bin/live_stream.rs
@@ -1,23 +1,23 @@
-/// Tool for users to stream footage from the cameras using Aravis and
-/// tune camera intrinsics quickly, rather than storing raw bytes to
-/// disk and converting to PNG afterhand.
-///
-/// This program requires 2 threads:
-///   (1) Main thread: UI rendering
-///   (2) Capture thread: Uses Aravis to get frames
-/// This is required for one, because capture calls `timeout_pop_buffer`
-/// which blocks the capture thread, and two, because macOS specifically
-/// requires a main thread, not a background thread, for windowing systems
-/// like the egui UI.
-use std::sync::{Arc, Mutex};
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::thread;
+// Tool for users to stream footage from the cameras using Aravis and
+// tune camera intrinsics quickly, rather than storing raw bytes to
+// disk and converting to PNG afterhand.
+//
+// This program requires 2 threads:
+//   (1) Main thread: UI rendering
+//   (2) Capture thread: Uses Aravis to get frames
+// This is required for one, because capture calls `timeout_pop_buffer`
+// which blocks the capture thread, and two, because macOS specifically
+// requires a main thread, not a background thread, for windowing systems
+// like the egui UI.
+use backend_lib::camera::CameraIngestConfig;
+use backend_lib::camera::StreamFromCamerasArgs;
 use clap::Parser;
 use eframe::egui;
-use backend_lib::camera::stream::{FrameData, LiveViewApp};
-use backend_lib::camera::stream::capture::run_capture_thread;
-use backend_lib::camera::CameraIngestConfig;
-use backend_lib::camera::stream::cli::StreamFromCamerasArgs;
+use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::{Arc, Mutex};
+use std::thread;
+use stream_lib::capture::run_capture_thread;
+use stream_lib::{FrameData, LiveViewApp};
 
 fn main() -> eframe::Result<()> {
     println!("------------------------");
@@ -26,7 +26,8 @@ fn main() -> eframe::Result<()> {
 
     // Parse command-line arguments and create camera configuration.
     let args: StreamFromCamerasArgs = StreamFromCamerasArgs::parse();
-    let camera_ingest_config: CameraIngestConfig = CameraIngestConfig::from_stream_args(args.clone());
+    let camera_ingest_config: CameraIngestConfig =
+        CameraIngestConfig::from_stream_args(args.clone());
 
     // Create shared camera ingest config object to be shared between the 2 threads.
     let camera_settings = Arc::new(Mutex::new(camera_ingest_config));
@@ -42,7 +43,8 @@ fn main() -> eframe::Result<()> {
     ctrlc::set_handler(move || {
         println!("\nShutdown signal received, stopping recording...");
         shutdown_clone.store(true, Ordering::SeqCst);
-    }).expect("Error setting Ctrl+C handler.");
+    })
+    .expect("Error setting Ctrl+C handler.");
 
     // Spawn capture thread.
     thread::spawn(move || {
@@ -54,7 +56,7 @@ fn main() -> eframe::Result<()> {
         // 983.0 came from 720.0 * 1.365, which is the division of 4096 / 3000 (camera sensor pixels).
         viewport: egui::ViewportBuilder::default()
             .with_title("Mjölnir Live Stream")
-            .with_inner_size([983.0, 720.0]), 
+            .with_inner_size([983.0, 720.0]),
         ..Default::default()
     };
