+++ backend/src/bin/record_from_both_cameras.rs
@@ -9,7 +9,7 @@ use clap::Parser;
 use backend_lib::camera::CameraIngestConfig;
 use backend_lib::camera::aravis_utils::initialize_aravis;
 use backend_lib::camera::discovery::get_camera_ids;
-use backend_lib::camera::record::cli::RecordWithBothCamerasArgs;
+use backend_lib::camera::RecordWithBothCamerasArgs;
 use backend_lib::camera::record::run_capture_thread;
 use backend_lib::camera::record::writer::{Frame, ensure_dir, write_to_disk};
