+++ backend/src/camera/cli_record.rs
