+++ backend/src/camera/cli_stream.rs
