+++ backend/src/camera/record/mod.rs
@@ -1,5 +1,4 @@
 pub mod capture;
-pub mod cli;
 pub mod writer;
 
 pub use capture::run_capture_thread;
