+++ backend/src/camera/stream/lib.rs
@@ -1,6 +1,5 @@
 pub mod app;
 pub mod capture;
-pub mod cli;
 pub mod frame;
 
 pub use app::LiveViewApp;
